:truck: move project finders function in dedicated file in environment
