+++ youwol/environment/projects_loader.py
@@ -1,20 +1,17 @@
 from __future__ import annotations
 
 import asyncio
-import itertools
 import os
 from pathlib import Path
 from pydantic import BaseModel
 from typing import List, Union, Optional, Awaitable, Iterable
 
-from youwol.configuration.defaults import default_path_projects_dir
+from youwol.environment.forward_declaration import YouwolEnvironment
 from youwol.environment.models import IPipelineFactory
 from youwol.environment.models_project import Project
-from youwol.environment.youwol_environment import YouwolEnvironment
 from youwol.utils.utils_low_level import get_object_from_module
 from youwol_utils import encode_id
 from youwol_utils.context import Context
-from youwol_utils.utils_paths import FileListing, matching_files
 
 PROJECT_PIPELINE_DIRECTORY = '.yw_pipeline'
 
@@ -179,36 +176,3 @@ async def get_project(project_path: Path,
             path=project_path
         )
 
-
-def default_projects_finder(env: YouwolEnvironment, root_folders: Union[None, str, Path, List[str], List[Path]] = None):
-    if not root_folders:
-        (Path.home() / default_path_projects_dir).mkdir(exist_ok=True)
-
-    root_folders = [Path.home() / default_path_projects_dir] if not root_folders else root_folders
-    root_folders = root_folders if isinstance(root_folders, List) else [root_folders]
-    results = [auto_detect_projects(env=env, root_folder=root_folder, ignore=["**/dist", '**/py-youwol'])
-               for root_folder in root_folders]
-
-    return itertools.chain.from_iterable(results)
-
-
-def auto_detect_projects(env: YouwolEnvironment, root_folder: Union[Path, str], ignore: List[str] = None):
-    database_ignore = None
-    system_ignore = None
-    root_folder = Path(root_folder)
-    try:
-        database_ignore = env.pathsBook.databases.relative_to(root_folder)
-    except ValueError:
-        pass
-    try:
-        system_ignore = env.pathsBook.system.relative_to(root_folder)
-    except ValueError:
-        pass
-
-    ignore = (ignore or []) + [str(path) for path in [database_ignore, system_ignore] if path]
-    file_listing = FileListing(
-        include=["**/.yw_pipeline/yw_pipeline.py"],
-        ignore=["**/node_modules", "**/.template", "**/.git"] + ignore
-    )
-    yw_pipelines = matching_files(root_folder, file_listing)
-    return [p.parent.parent for p in yw_pipelines]
